nix automation to run integration tests
